requirements.txt
transformers
torch
sentence-transformers
chromadb
duckduckgo-search
fastapi
uvicorn
requests
jinja2
openai-whisper
